cleanup old wasm api
