+++ packages/lib-wasm/src/processing/mod.rs
@@ -8,8 +8,4 @@ pub mod navigator;
 
 pub mod sds_processor;
 
-pub mod sds_processor_v2;
-
 pub mod sds_context;
-
-pub mod sds_context_v2;
