+++ packages/lib-wasm/src/processing/navigator/navigate_result.rs
@@ -14,10 +14,7 @@ use std::{cmp::Reverse, convert::TryFrom, sync::Arc};
 use wasm_bindgen::prelude::*;
 
 use crate::{
-    processing::{
-        aggregator::WasmAggregateResult, sds_processor::SDSProcessor,
-        sds_processor_v2::WasmSdsProcessor,
-    },
+    processing::{aggregator::WasmAggregateResult, sds_processor::WasmSdsProcessor},
     utils::js::{
         JsAttributesIntersectionByColumn, JsHeaderNames, JsResult, JsSelectedAttributesByColumn,
     },
@@ -138,24 +135,8 @@ impl WasmNavigateResult {
 
 #[wasm_bindgen]
 impl WasmNavigateResult {
-    // TODO: remove this
     #[wasm_bindgen(constructor)]
-    pub fn from_synthetic_processor(synthetic_processor: &SDSProcessor) -> WasmNavigateResult {
-        WasmNavigateResult::new(
-            synthetic_processor.data_block.clone(),
-            synthetic_processor
-                .data_block
-                .calc_attr_rows_by_column_with_no_empty_values(),
-            WasmSelectedAttributesByColumn::default(),
-            AttributeRows::default(),
-            (0..synthetic_processor.data_block.number_of_records()).collect(),
-            synthetic_processor.data_block.calc_column_index_by_name(),
-        )
-    }
-
-    pub fn from_synthetic_processor_v2(
-        synthetic_processor: &WasmSdsProcessor,
-    ) -> WasmNavigateResult {
+    pub fn from_synthetic_processor(synthetic_processor: &WasmSdsProcessor) -> WasmNavigateResult {
         WasmNavigateResult::new(
             synthetic_processor.data_block.clone(),
             synthetic_processor
