+++ packages/lib-wasm/src/processing/sds_processor/base_synthesis_parameters.rs
