+++ packages/lib-wasm/src/processing/sds_processor/mod.rs
@@ -1,7 +1,15 @@
+mod base_synthesis_parameters;
+mod csv_data_parameters;
+mod dp_parameters;
 mod header_names;
-mod input_number_by_length;
+mod noisy_count_threshold;
+mod oversampling_parameters;
 mod processor;
 
+pub use base_synthesis_parameters::*;
+pub use csv_data_parameters::*;
+pub use dp_parameters::*;
 pub use header_names::*;
-pub use input_number_by_length::*;
+pub use noisy_count_threshold::*;
+pub use oversampling_parameters::*;
 pub use processor::*;
