+++ packages/lib-wasm/src/processing/sds_processor/processor.rs
@@ -1,197 +1,227 @@
+use super::{WasmBaseSynthesisParameters, WasmOversamplingParameters};
 use csv::ReaderBuilder;
 use js_sys::Function;
 use sds_core::{
-    data_block::{CsvDataBlockCreator, CsvRecord, DataBlock, DataBlockCreator},
-    dp::{DpParameters, InputValueByLen, NoisyCountThreshold},
+    data_block::{CsvDataBlockCreator, DataBlock, DataBlockCreator},
+    dp::{DpParameters, NoisyCountThreshold},
     processing::{
         aggregator::Aggregator,
         generator::{Generator, OversamplingParameters},
     },
     utils::time::ElapsedDurationLogger,
 };
+use std::io::Cursor;
 use std::sync::Arc;
-use std::{convert::TryFrom, io::Cursor};
 use wasm_bindgen::{prelude::*, JsCast};
 
 use crate::{
-    utils::js::{JsInputNumberByLength, JsProgressReporter},
-    {
-        processing::{aggregator::WasmAggregateResult, generator::WasmGenerateResult},
-        utils::js::{JsHeaderNames, JsReportProgressCallback},
-    },
+    processing::{generator::WasmGenerateResult, sds_processor::WasmCsvDataParameters},
+    utils::js::{JsDpParameters, JsNoisyCountThreshold, JsProgressReporter, JsResult},
+    {processing::aggregator::WasmAggregateResult, utils::js::JsReportProgressCallback},
 };
 
+const DEFAULT_CACHE_MAX_SIZE: usize = 100000;
+
+const DEFAULT_EMPTY_VALUE: &str = "";
+
 #[wasm_bindgen]
-pub struct SDSProcessor {
+pub struct WasmSdsProcessor {
     pub(crate) data_block: Arc<DataBlock>,
 }
 
-#[wasm_bindgen]
-impl SDSProcessor {
-    pub fn default() -> SDSProcessor {
-        SDSProcessor {
+impl WasmSdsProcessor {
+    pub fn default() -> WasmSdsProcessor {
+        WasmSdsProcessor {
             data_block: Arc::new(DataBlock::default()),
         }
     }
 }
 
-#[wasm_bindgen]
-impl SDSProcessor {
+#[wasm_bindgen(constructor)]
+impl WasmSdsProcessor {
+    #[inline]
     #[wasm_bindgen(constructor)]
     pub fn new(
         csv_data: &str,
-        delimiter: char,
-        use_columns: JsHeaderNames,
-        sensitive_zeros: JsHeaderNames,
-        record_limit: usize,
-    ) -> Result<SDSProcessor, JsValue> {
-        let _duration_logger = ElapsedDurationLogger::new(String::from("sds processor creation"));
+        csv_data_params: &WasmCsvDataParameters,
+    ) -> JsResult<WasmSdsProcessor> {
+        let _duration_logger = ElapsedDurationLogger::new("sds processor creation");
         let data_block = CsvDataBlockCreator::create(
             Ok(ReaderBuilder::new()
-                .delimiter(delimiter as u8)
+                .delimiter(csv_data_params.delimiter as u8)
                 .from_reader(Cursor::new(csv_data))),
-            &CsvRecord::try_from(use_columns)?,
-            &CsvRecord::try_from(sensitive_zeros)?,
-            record_limit,
+            &csv_data_params.use_columns,
+            &csv_data_params.sensitive_zeros,
+            csv_data_params.record_limit,
         )
         .map_err(|err| JsValue::from(err.to_string()))?;
 
-        Ok(SDSProcessor { data_block })
+        Ok(WasmSdsProcessor { data_block })
     }
 
-    #[wasm_bindgen(js_name = "numberOfRecords")]
-    pub fn number_of_records(&self) -> usize {
-        self.data_block.number_of_records()
-    }
-
-    #[wasm_bindgen(js_name = "protectedNumberOfRecords")]
-    pub fn protected_number_of_records(&self, resolution: usize) -> usize {
-        self.data_block.protected_number_of_records(resolution)
-    }
-
-    #[wasm_bindgen(js_name = "normalizeReportingLength")]
-    pub fn normalize_reporting_length(&self, reporting_length: usize) -> usize {
-        self.data_block.normalize_reporting_length(reporting_length)
-    }
-
+    #[inline]
+    #[wasm_bindgen(js_name = "aggregate")]
     pub fn aggregate(
         &self,
         reporting_length: usize,
         progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmAggregateResult, JsValue> {
-        let _duration_logger =
-            ElapsedDurationLogger::new(String::from("sds processor aggregation"));
+    ) -> JsResult<WasmAggregateResult> {
         let js_callback: Function = progress_callback.dyn_into()?;
-        let mut aggregator = Aggregator::new(self.data_block.clone());
 
-        aggregator
-            .aggregate(
-                reporting_length,
-                &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
-            )
+        self._aggregate(
+            reporting_length,
+            &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+        )
+    }
+
+    #[inline]
+    #[wasm_bindgen(js_name = "aggregateWithDp")]
+    pub fn aggregate_with_dp(
+        &self,
+        reporting_length: usize,
+        dp_parameters: JsDpParameters,
+        threshold: JsNoisyCountThreshold,
+        progress_callback: JsReportProgressCallback,
+    ) -> JsResult<WasmAggregateResult> {
+        let js_callback: Function = progress_callback.dyn_into()?;
+
+        self._aggregate_with_dp(
+            reporting_length,
+            dp_parameters,
+            threshold,
+            &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+        )
+    }
+
+    #[inline]
+    #[wasm_bindgen(js_name = "generateUnseeded")]
+    pub fn generate_unseeded(
+        &self,
+        base_parameters: &WasmBaseSynthesisParameters,
+        progress_callback: JsReportProgressCallback,
+    ) -> JsResult<WasmGenerateResult> {
+        let js_callback: Function = progress_callback.dyn_into()?;
+
+        self._generate_unseeded(
+            base_parameters,
+            &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+        )
+    }
+
+    #[inline]
+    #[wasm_bindgen(js_name = "generateRowSeeded")]
+    pub fn generate_row_seeded(
+        &self,
+        base_parameters: &WasmBaseSynthesisParameters,
+        progress_callback: JsReportProgressCallback,
+    ) -> JsResult<WasmGenerateResult> {
+        let js_callback: Function = progress_callback.dyn_into()?;
+
+        self._generate_row_seeded(
+            base_parameters,
+            &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+        )
+    }
+
+    #[inline]
+    #[wasm_bindgen(js_name = "generateValueSeeded")]
+    pub fn generate_value_seeded(
+        &self,
+        base_parameters: &WasmBaseSynthesisParameters,
+        aggregated_result: &WasmAggregateResult,
+        oversampling_parameters: Option<WasmOversamplingParameters>,
+        progress_callback: JsReportProgressCallback,
+    ) -> JsResult<WasmGenerateResult> {
+        let js_callback: Function = progress_callback.dyn_into()?;
+
+        self._generate_value_seeded(
+            base_parameters,
+            aggregated_result,
+            oversampling_parameters,
+            &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+        )
+    }
+
+    #[inline]
+    #[wasm_bindgen(js_name = "generateAggregateSeeded")]
+    pub fn generate_aggregate_seeded(
+        &self,
+        base_parameters: &WasmBaseSynthesisParameters,
+        aggregated_result: &WasmAggregateResult,
+        use_synthetic_counts: bool,
+        progress_callback: JsReportProgressCallback,
+    ) -> JsResult<WasmGenerateResult> {
+        let js_callback: Function = progress_callback.dyn_into()?;
+
+        self._generate_aggregate_seeded(
+            base_parameters,
+            aggregated_result,
+            use_synthetic_counts,
+            &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+        )
+    }
+}
+
+impl WasmSdsProcessor {
+    #[inline]
+    fn unwrap_base_synthesis_parameters_or_default(
+        base_parameters: &WasmBaseSynthesisParameters,
+    ) -> (usize, usize, String) {
+        (
+            base_parameters.resolution,
+            base_parameters
+                .cache_max_size
+                .unwrap_or(DEFAULT_CACHE_MAX_SIZE),
+            base_parameters
+                .empty_value
+                .clone()
+                .unwrap_or_else(|| DEFAULT_EMPTY_VALUE.to_owned()),
+        )
+    }
+
+    #[inline]
+    pub(crate) fn _aggregate(
+        &self,
+        reporting_length: usize,
+        progress_reporter: &mut Option<JsProgressReporter>,
+    ) -> JsResult<WasmAggregateResult> {
+        Aggregator::new(self.data_block.clone())
+            .aggregate(reporting_length, progress_reporter)
             .map(|aggregated_data| WasmAggregateResult::new(Arc::new(aggregated_data)))
             .map_err(|err| JsValue::from(err.to_string()))
     }
 
     #[inline]
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn aggregate_with_dp(
+    pub(crate) fn _aggregate_with_dp(
         &self,
         reporting_length: usize,
-        epsilon: f64,
-        delta: f64,
-        percentile_percentage: usize,
-        percentile_epsilon_proportion: f64,
-        sigma_proportions: Option<Vec<f64>>,
-        threshold: NoisyCountThreshold,
-        progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmAggregateResult, JsValue> {
-        let _duration_logger = ElapsedDurationLogger::new(format!(
-            "sds processor aggregation with dp and threshold = {:?}",
-            threshold
-        ));
-        let js_callback: Function = progress_callback.dyn_into()?;
+        dp_parameters: JsDpParameters,
+        threshold: JsNoisyCountThreshold,
+        progress_reporter: &mut Option<JsProgressReporter>,
+    ) -> JsResult<WasmAggregateResult> {
         let aggregator = Aggregator::new(self.data_block.clone());
 
         Ok(WasmAggregateResult::new(Arc::new(
             aggregator
                 .aggregate_with_dp(
                     reporting_length,
-                    &DpParameters::new(
-                        epsilon,
-                        delta,
-                        percentile_percentage,
-                        percentile_epsilon_proportion,
-                        sigma_proportions,
-                    ),
-                    threshold,
-                    &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+                    &DpParameters::try_from(dp_parameters)?,
+                    NoisyCountThreshold::try_from(threshold)?,
+                    progress_reporter,
                 )
                 .map_err(|err| JsValue::from(err.to_string()))?,
         )))
     }
 
-    #[wasm_bindgen(js_name = "aggregateWithDpFixedThreshold")]
-    #[allow(clippy::too_many_arguments)]
-    pub fn aggregate_with_dp_fixed_threshold(
+    #[inline]
+    pub(crate) fn _generate_unseeded(
         &self,
-        reporting_length: usize,
-        epsilon: f64,
-        delta: f64,
-        percentile_percentage: usize,
-        percentile_epsilon_proportion: f64,
-        sigma_proportions: Option<Vec<f64>>,
-        threshold: JsInputNumberByLength,
-        progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmAggregateResult, JsValue> {
-        self.aggregate_with_dp(
-            reporting_length,
-            epsilon,
-            delta,
-            percentile_percentage,
-            percentile_epsilon_proportion,
-            sigma_proportions,
-            NoisyCountThreshold::Fixed(InputValueByLen::<f64>::try_from(threshold)?),
-            progress_callback,
-        )
-    }
-
-    #[wasm_bindgen(js_name = "aggregateWithDpAdaptiveThreshold")]
-    #[allow(clippy::too_many_arguments)]
-    pub fn aggregate_with_dp_adaptive_threshold(
-        &self,
-        reporting_length: usize,
-        epsilon: f64,
-        delta: f64,
-        percentile_percentage: usize,
-        percentile_epsilon_proportion: f64,
-        sigma_proportions: Option<Vec<f64>>,
-        threshold: JsInputNumberByLength,
-        progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmAggregateResult, JsValue> {
-        self.aggregate_with_dp(
-            reporting_length,
-            epsilon,
-            delta,
-            percentile_percentage,
-            percentile_epsilon_proportion,
-            sigma_proportions,
-            NoisyCountThreshold::Adaptive(InputValueByLen::<f64>::try_from(threshold)?),
-            progress_callback,
-        )
-    }
-
-    #[wasm_bindgen(js_name = "generateUnseeded")]
-    pub fn generate_unseeded(
-        &self,
-        cache_max_size: usize,
-        resolution: usize,
-        empty_value: String,
-        progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmGenerateResult, JsValue> {
-        let _duration_logger = ElapsedDurationLogger::new(String::from("generation unseeded"));
-        let js_callback: Function = progress_callback.dyn_into()?;
+        base_parameters: &WasmBaseSynthesisParameters,
+        progress_reporter: &mut Option<JsProgressReporter>,
+    ) -> JsResult<WasmGenerateResult> {
         let generator = Generator::default();
+        let (resolution, cache_max_size, empty_value) =
+            WasmSdsProcessor::unwrap_base_synthesis_parameters_or_default(base_parameters);
 
         Ok(WasmGenerateResult::new(
             generator.generate_unseeded(
@@ -199,23 +229,21 @@ impl SDSProcessor {
                 resolution,
                 cache_max_size,
                 &empty_value,
-                &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+                progress_reporter,
             ),
             resolution,
         ))
     }
 
-    #[wasm_bindgen(js_name = "generateRowSeeded")]
-    pub fn generate_row_seeded(
+    #[inline]
+    pub(crate) fn _generate_row_seeded(
         &self,
-        cache_max_size: usize,
-        resolution: usize,
-        empty_value: String,
-        progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmGenerateResult, JsValue> {
-        let _duration_logger = ElapsedDurationLogger::new(String::from("generation row seeded"));
-        let js_callback: Function = progress_callback.dyn_into()?;
+        base_parameters: &WasmBaseSynthesisParameters,
+        progress_reporter: &mut Option<JsProgressReporter>,
+    ) -> JsResult<WasmGenerateResult> {
         let generator = Generator::default();
+        let (resolution, cache_max_size, empty_value) =
+            WasmSdsProcessor::unwrap_base_synthesis_parameters_or_default(base_parameters);
 
         Ok(WasmGenerateResult::new(
             generator.generate_row_seeded(
@@ -223,27 +251,23 @@ impl SDSProcessor {
                 resolution,
                 cache_max_size,
                 &empty_value,
-                &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+                progress_reporter,
             ),
             resolution,
         ))
     }
 
-    #[wasm_bindgen(js_name = "generateValueSeeded")]
-    #[allow(clippy::too_many_arguments)]
-    pub fn generate_value_seeded(
+    #[inline]
+    pub(crate) fn _generate_value_seeded(
         &self,
-        cache_max_size: usize,
-        resolution: usize,
-        empty_value: String,
+        base_parameters: &WasmBaseSynthesisParameters,
         aggregated_result: &WasmAggregateResult,
-        oversampling_ratio: Option<f64>,
-        oversampling_tries: Option<usize>,
-        progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmGenerateResult, JsValue> {
-        let _duration_logger = ElapsedDurationLogger::new(String::from("generation value seeded"));
-        let js_callback: Function = progress_callback.dyn_into()?;
+        oversampling_parameters: Option<WasmOversamplingParameters>,
+        progress_reporter: &mut Option<JsProgressReporter>,
+    ) -> JsResult<WasmGenerateResult> {
         let generator = Generator::default();
+        let (resolution, cache_max_size, empty_value) =
+            WasmSdsProcessor::unwrap_base_synthesis_parameters_or_default(base_parameters);
 
         Ok(WasmGenerateResult::new(
             generator.generate_value_seeded(
@@ -251,38 +275,39 @@ impl SDSProcessor {
                 resolution,
                 cache_max_size,
                 &empty_value,
-                oversampling_ratio.map(|_| OversamplingParameters {
-                    aggregated_data: aggregated_result.aggregated_data.clone(),
-                    oversampling_ratio,
-                    oversampling_tries,
+                oversampling_parameters.map(|params| {
+                    OversamplingParameters::new(
+                        aggregated_result.aggregated_data.clone(),
+                        params.oversampling_ratio,
+                        params.oversampling_tries,
+                    )
                 }),
-                &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+                progress_reporter,
             ),
             resolution,
         ))
     }
 
-    #[wasm_bindgen(js_name = "generateAggregateSeeded")]
-    pub fn generate_aggregate_seeded(
+    #[inline]
+    pub(crate) fn _generate_aggregate_seeded(
         &self,
-        empty_value: String,
+        base_parameters: &WasmBaseSynthesisParameters,
         aggregated_result: &WasmAggregateResult,
         use_synthetic_counts: bool,
-        progress_callback: JsReportProgressCallback,
-    ) -> Result<WasmGenerateResult, JsValue> {
-        let _duration_logger =
-            ElapsedDurationLogger::new(String::from("generation aggregate seeded"));
-        let js_callback: Function = progress_callback.dyn_into()?;
+        progress_reporter: &mut Option<JsProgressReporter>,
+    ) -> JsResult<WasmGenerateResult> {
         let generator = Generator::default();
+        let (resolution, _, empty_value) =
+            WasmSdsProcessor::unwrap_base_synthesis_parameters_or_default(base_parameters);
 
         Ok(WasmGenerateResult::new(
             generator.generate_aggregate_seeded(
                 &empty_value,
                 aggregated_result.aggregated_data.clone(),
                 use_synthetic_counts,
-                &mut Some(JsProgressReporter::new(&js_callback, &|p| p)),
+                progress_reporter,
             ),
-            1,
+            resolution,
         ))
     }
 }
