--- packages/lib-wasm/src/processing/sds_processor_v2/mod.rs
@@ -1,15 +1,0 @@
-mod base_synthesis_parameters;
-mod csv_data_parameters;
-mod dp_parameters;
-mod header_names;
-mod noisy_count_threshold;
-mod oversampling_parameters;
-mod processor;
-
-pub use base_synthesis_parameters::*;
-pub use csv_data_parameters::*;
-pub use dp_parameters::*;
-pub use header_names::*;
-pub use noisy_count_threshold::*;
-pub use oversampling_parameters::*;
-pub use processor::*;
